+++ CHANGES.md
@@ -1,8 +1,15 @@
 # CHANGES
 
-## v3.6.9-RC
+## v3.6.10
 
 *
+
+## v3.6.9
+
+* Improvement: Redirection when login/logout
+* Improvement: Add home icon before '/'
+* Fix: Client crashed when the first login
+    * Introduced by 3.6.8
 
 ## v3.6.8
 
@@ -10,6 +17,7 @@
 * Improvement: Optimize markdown rendering
 * Improvement: Reactify admin pages (Navigation)
 * Fix: Reply comments collapsed are broken
+    * Introduced by 3.6.7
 * Support: Update libs
     * cross-env
     * mkdirp